Refactor `ResetPasswordController` by introducing `ResetPasswordAction` and `SendResetLinkAction` for cleaner logic separation.
