+++ app/Http/Controllers/ResetPasswordController.php
@@ -4,49 +4,28 @@ declare(strict_types=1);
 
 namespace Interns2025b\Http\Controllers;
 
-use Illuminate\Auth\Events\PasswordReset;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Password;
-use Illuminate\Support\Str;
+use Interns2025b\Actions\ResetPasswordAction;
+use Interns2025b\Actions\SendResetLinkAction;
 use Interns2025b\Http\Requests\PasswordResetLinkRequest;
 use Interns2025b\Http\Requests\ResetPasswordRequest;
-use Interns2025b\Models\User;
 use Symfony\Component\HttpFoundation\Response as Status;
 
 class ResetPasswordController extends Controller
 {
-    public function sendResetLinkEmail(PasswordResetLinkRequest $request): JsonResponse
+    public function sendResetLinkEmail(PasswordResetLinkRequest $request, SendResetLinkAction $action): JsonResponse
     {
-        $validated = $request->validated();
-        Password::sendResetLink($validated);
-
-        activity()
-            ->withProperties(["email" => $request->input("email")])
-            ->log("Requested password reset link via API");
+        $message = $action->execute($request->validated());
 
         return response()->json([
-            "message" => __("passwords.sent"),
+            "message" => $message,
         ], Status::HTTP_OK);
     }
 
-    public function resetPassword(ResetPasswordRequest $request): JsonResponse
+    public function resetPassword(ResetPasswordRequest $request, ResetPasswordAction $action): JsonResponse
     {
-        $status = Password::reset(
-            $request->only("email", "password", "password_confirmation", "token"),
-            function (User $user, string $password): void {
-                $user->forceFill([
-                    "password" => Hash::make($password),
-                    "remember_token" => Str::random(60),
-                ])->save();
-
-                event(new PasswordReset($user));
-
-                activity()
-                    ->performedOn($user)
-                    ->log("Reset password via API");
-            },
-        );
+        $status = $action->execute($request->only("email", "password", "password_confirmation", "token"));
 
         return $status === Password::PASSWORD_RESET
             ? response()->json(["message" => __("passwords.reset")], Status::HTTP_OK)
